Make update_archive.py be a wrapper script to a package module
